requirements.txt
streamlit
requests
Pillow
streamlit_option_menu
